backend/requirements.txt
fastapi
uvicorn
pandas
python-dotenv
sentence-transformers
faiss-cpu
requests
